packages/si-agenda/pyproject.toml
[project]
name = "si-agenda"
version = "0.1.0"
description = "SI Ecoles Agenda Scraper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer>=0.9.0",
    "ics>=0.7.2",
    "pydantic>=2.5.0",
    "cas-connector",
]

[project.scripts]
si-agenda = "si_agenda.cli:app"

[tool.uv.sources]
cas-connector = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/si_agenda"]
